+++ src/api/firebase.ts
@@ -1,52 +1,22 @@
 import { FirebaseApp, initializeApp } from 'firebase/app';
 import { getFirestore } from 'firebase/firestore';
 import { getStorage } from 'firebase/storage';
-import {
-    UserCredential,
-    signInWithEmailAndPassword,
-    signOut,
-    onAuthStateChanged,
-    NextOrObserver,
-    User,
-    getAuth,
-    sendPasswordResetEmail,
-    signInAnonymously
-} from 'firebase/auth';
+import { User, getAuth } from 'firebase/auth';
 
 import { firebaseConfig } from './config';
-import {
-    addEvent,
-    checkClaims,
-    getImageAsSignedUrl,
-    getUidByEmail,
-    registerNewUser,
-    setClaimForAdmin,
-    setClaimForAidWorker,
-    setClaimForDonationReadAccess,
-    setClaimForDonor,
-    setClaimForNewUser,
-    setClaimForVerified,
-    setClaimForVolunteer,
-    setClaims,
-    setUserAccount,
-    toggleCanReadDonations,
-    toggleClaimForAdmin,
-    toggleClaimForAidWorker,
-    toggleClaimForVerified,
-    toggleClaimForVolunteer,
-    getAuthUserById
-} from './firebaseAdmin';
+import { addEvent, checkClaims } from './firebaseAdmin';
 
 import { getFunctions, httpsCallable } from 'firebase/functions';
 
-import { AccountInformation } from '@/types/UserTypes';
+import { AccountInformation, NewUserAccountInfo, AuthUserRecord } from '@/types/UserTypes';
 import { convertToString } from '@/utils/utils';
-import { AuthUserRecord, newUserAccountInfo } from '@/types/UserTypes';
-import { Auth, UserRecord } from 'firebase-admin/auth';
-import { IUser } from '@/models/user';
-import { OrganizationNames } from '@/types/OrganizationTypes';
+import { UserRecord } from 'firebase-admin/auth';
 
 export const app: FirebaseApp = initializeApp(firebaseConfig);
+
+export const db = getFirestore(app);
+export const storage = getStorage(app);
+export const auth = getAuth(app);
 
 //Cloud functions
 const functions = getFunctions(app);
@@ -58,9 +28,16 @@ const listAllUsers = httpsCallable(functions, 'listallusers');
 const setCustomClaims = httpsCallable(functions, 'setcustomclaims');
 const updateAuthUser = httpsCallable(functions, 'updateauthuser');
 
-export const db = getFirestore(app);
-export const storage = getStorage(app);
-export const auth = getAuth(app);
+//Cloud function calls
+export async function callCreateUser(accountInfo: NewUserAccountInfo): Promise<UserRecord> {
+    try {
+        const result = await createNewUser(accountInfo);
+        return result.data as UserRecord;
+    } catch (error) {
+        addErrorEvent('Create user', error);
+    }
+    return Promise.reject();
+}
 
 export async function callCheckClaims(...claimNames: string[]): Promise<any> {
     if (claimNames.length === 0) {
@@ -108,6 +85,15 @@ export async function callUpdateAuthUser(uid: string, accountInformation: Accoun
     return Promise.reject();
 }
 
+export async function callEnableUser(userId: string): Promise<void> {
+    try {
+        await enableUser({ userId: userId });
+    } catch (error) {
+        addErrorEvent('Could not enable user', error);
+    }
+    return Promise.reject();
+}
+
 export async function callSetClaims(userId: string, claims: any): Promise<void> {
     try {
         await setCustomClaims({ userId: userId, claims: claims });
@@ -116,17 +102,18 @@ export async function callSetClaims(userId: string, claims: any): Promise<void> 
     }
 }
 
-// export async function callSetClaimForNewUser(userId: string): Promise<void> {
-//     setClaimForNewUser({ userId: userId });
-// }
-
-// Action based claims.
-export async function callSetClaimForDonationReadAccess(userId: string, canReadDonations: boolean): Promise<void> {
-    setClaimForDonationReadAccess({ userId: userId, canReadDonations: canReadDonations });
-}
-
-export async function callToggleCanReadDonations(userId: string): Promise<void> {
-    toggleCanReadDonations({ userId: userId });
+export async function callGetOrganizationNames(): Promise<{
+    [key: string]: string;
+}> {
+    try {
+        const orgNames = await getOrganizationNames();
+        return orgNames.data as {
+            [key: string]: string;
+        };
+    } catch (error) {
+        addErrorEvent('Could not fetch organization names', error);
+    }
+    return Promise.reject();
 }
 
 // Role based claims.
@@ -153,116 +140,4 @@ export async function checkIsAidWorker(user: User): Promise<boolean> {
 // Utilitarian
 export async function addErrorEvent(location: string, error: any): Promise<void> {
     addEvent({ location: location, error: convertToString(error) });
-}
-
-export async function callGetImageAsSignedUrl(url: string): Promise<any> {
-    try {
-        const response = await getImageAsSignedUrl({ url: url });
-        const signedUrl: any = response.data;
-        return signedUrl;
-    } catch (error) {
-        await addErrorEvent('getImageAsSignedUrl', error);
-    }
-}
-
-export function getUserEmail(): string | null | undefined {
-    return auth.currentUser?.email;
-}
-
-export async function callGetUidByEmail(email: string): Promise<string> {
-    await auth.authStateReady();
-    const options = { email };
-    const uid = await getUidByEmail({ options });
-    return uid;
-}
-
-export async function getUserId(): Promise<string> {
-    await auth.authStateReady();
-    const currentUser = auth.currentUser?.uid;
-    return currentUser ?? Promise.reject();
-}
-
-export async function getUserEmailById(id: string): Promise<string> {
-    try {
-        const user = await getAuthUserById(id);
-        if (user.email) return user.email;
-    } catch (error) {
-        addErrorEvent('Get user email by ID', error);
-    }
-    return Promise.reject();
-}
-
-export async function callRegisterNewUser(options: any): Promise<any> {
-    const response = await registerNewUser(options);
-    const data: any = response.data;
-    const okResponse = data.ok ?? false;
-    return { ok: okResponse };
-}
-
-export async function callSetUserAccount(userId: string, accountInformation: AccountInformation): Promise<void> {
-    await setUserAccount({ userId: userId, accountInformation: accountInformation });
-}
-
-export async function createUser(accountInfo: newUserAccountInfo): Promise<UserRecord> {
-    try {
-        const result = await createNewUser(accountInfo);
-        return result.data as UserRecord;
-    } catch (error) {
-        addErrorEvent('Create user', error);
-    }
-    return Promise.reject();
-}
-
-export async function callEnableUser(userId: string): Promise<void> {
-    try {
-        await enableUser({ userId: userId });
-    } catch (error) {
-        addErrorEvent('Could not enable user', error);
-    }
-    return Promise.reject();
-}
-
-export async function callGetOrganizationNames(): Promise<{
-    [key: string]: string;
-}> {
-    try {
-        const orgNames = await getOrganizationNames();
-        return orgNames.data as {
-            [key: string]: string;
-        };
-    } catch (error) {
-        addErrorEvent('Could not fetch organization names', error);
-    }
-    return Promise.reject();
-}
-
-export async function signInAuthUserWithEmailAndPassword(email: string, password: string): Promise<null | User> {
-    if (!email || !password) {
-        return null;
-    }
-    const userCredential: UserCredential = await signInWithEmailAndPassword(auth, email, password);
-    return userCredential.user;
-}
-
-export function signOutUser(): void {
-    signOut(auth);
-}
-
-export function onAuthStateChangedListener(callback: NextOrObserver<User>) {
-    onAuthStateChanged(auth, callback);
-}
-
-export async function resetPassword(email: string): Promise<void> {
-    if (!email) Promise.reject();
-    await sendPasswordResetEmail(auth, email);
-}
-
-export async function loginAnonymousUser(): Promise<User | null> {
-    try {
-        const userCredential = await signInAnonymously(auth);
-        return userCredential.user;
-    } catch (error) {
-        addErrorEvent('Login anonymously', error);
-    }
-    return Promise.reject();
 }
